requirements.txt
annotated-types==0.7.0
anthropic==0.34.1
anyio==4.4.0
azure-cognitiveservices-search-websearch==2.0.1
azure-common==1.1.28
azure-core==1.30.2
azure-mgmt-core==1.4.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
filelock==3.15.4
fsspec==2024.6.1
google_search_results==2.4.2
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
isodate==0.6.1
jiter==0.5.0
msrest==0.7.1
numpy==2.0.1
oauthlib==3.2.2
ollama==0.3.2
openai==1.40.6
packaging==24.1
pandas==2.2.2
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
serpapi==0.1.5
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tiktoken==0.7.0
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
